requirements.txt
fastapi
uvicorn
python-socketio
google-generativeai
python-dotenv
scapy
netifaces
requests
tqdm